Add functions to global_det_atmos_util
Refs: #1
